split movement calculations into module
